Routes: Rename internship to internship-su24
TEST=manual
