+++ js/src/patina/pages/past-programs/internship-su24/FractalBadge.tsx
