+++ js/src/patina/pages/past-programs/internship-su24/InternCard.tsx
